Revert "Merge branch 'feature/ps-1995-event-resource-plugins' into feature/ps-1131-pub-and-sync-resync"
This reverts commit 0f0dbdbff910e711ddf701bfda6945f12bb78ba3, reversing
changes made to 739a7ffe7468e67320a9d5991e6783611a57cf4f.
