+++ src/Spryker/Zed/EventBehavior/Business/EventBehaviorBusinessFactory.php
@@ -9,9 +9,6 @@ namespace Spryker\Zed\EventBehavior\Business;
 
 use Spryker\Zed\EventBehavior\Business\Model\EventEntityTransferFilter;
 use Spryker\Zed\EventBehavior\Business\Model\EventResourceManager;
-use Spryker\Zed\EventBehavior\Business\Model\EventResourcePluginResolver;
-use Spryker\Zed\EventBehavior\Business\Model\EventResourceQueryContainerManager;
-use Spryker\Zed\EventBehavior\Business\Model\EventResourceRepositoryManager;
 use Spryker\Zed\EventBehavior\Business\Model\TriggerManager;
 use Spryker\Zed\EventBehavior\EventBehaviorDependencyProvider;
 use Spryker\Zed\Kernel\Business\AbstractBusinessFactory;
@@ -36,32 +33,12 @@ class EventBehaviorBusinessFactory extends AbstractBusinessFactory
     }
 
     /**
-     * @return \Spryker\Zed\EventBehavior\Business\Model\EventResourceQueryContainerManager
+     * @return \Spryker\Zed\EventBehavior\Business\Model\EventResourceManagerInterface
      */
-    public function createEventResourceQueryContainerManager()
+    public function createEventResourceManager()
     {
-        return new EventResourceQueryContainerManager(
+        return new EventResourceManager(
             $this->getEventFacade(),
-            $this->getEventResourcePlugins()
-        );
-    }
-
-    /**
-     * @return \Spryker\Zed\EventBehavior\Business\Model\EventResourceRepositoryManager
-     */
-    public function createEventResourceRepositoryManager()
-    {
-        return new EventResourceRepositoryManager(
-            $this->getEventFacade(),
-            $this->getEventResourcePlugins()
-        );
-    }
-
-    public function createEventResourcePluginResolver()
-    {
-        return new EventResourcePluginResolver(
-            $this->createEventResourceRepositoryManager(),
-            $this->createEventResourceQueryContainerManager(),
             $this->getEventResourcePlugins()
         );
     }
