+++ src/Spryker/Zed/EventBehavior/Business/EventBehaviorFacadeInterface.php
@@ -82,5 +82,5 @@ interface EventBehaviorFacadeInterface
      *
      * @return void
      */
-    public function executeResolvedPluginsBySources(array $resources, array $ids = []);
+    public function triggerResourceEvents(array $resources, array $ids = []);
 }
