--- src/Spryker/Zed/EventBehavior/Dependency/Plugin/EventResourceRepositoryPluginInterface.php
@@ -1,22 +1,0 @@
-<?php
-
-/**
- * Copyright © 2016-present Spryker Systems GmbH. All rights reserved.
- * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
- */
-
-namespace Spryker\Zed\EventBehavior\Dependency\Plugin;
-
-interface EventResourceRepositoryPluginInterface extends EventResourcePluginInterface
-{
-    /**
-     * Specification:
-     *  - Returns query of resource entity, provided $ids parameter
-     *    will apply to query to limit the result
-     *
-     * @api
-     *
-     * @return \Spryker\Shared\Kernel\Transfer\AbstractEntityTransfer[]
-     */
-    public function getData();
-}
